+++ client/src/app/app.routes.ts
@@ -62,8 +62,8 @@ export const appRoutes: Route[] = [
             initialData: initialDataResolver
         },
         children: [
-            {path: 'dashboard', loadChildren: () => import('app/modules/admin/example/example.routes')},
-            {path: 'curriculum', loadChildren: () => import('app/modules/admin/curriculum/curriculum.routes')},
+            {path: 'dashboard', loadChildren: () => import('app/modules/superadmin/example/example.routes')},
+            {path: 'curriculum', loadChildren: () => import('app/modules/superadmin/curriculum/curriculum.routes')},
         ]
     }
 ];
