+++ client/src/app/modules/superadmin/curriculum/curriculum-list/curriculum-list.component.spec.ts
