+++ client/src/app/modules/superadmin/curriculum/curriculum-list/curriculum.service.ts
