+++ client/src/app/modules/superadmin/curriculum/curriculum.routes.ts
