+++ client/src/app/modules/superadmin/curriculum/curriculum.types.ts
