+++ client/src/app/state/curriculum/curriculum.reducer.ts
@@ -1,6 +1,6 @@
 import { createReducer, on } from '@ngrx/store';
 import * as CurriculumActions from './curriculum.actions';
-import { ICurriculum } from 'app/modules/admin/curriculum/curriculum.types';
+import { ICurriculum } from 'app/modules/superadmin/curriculum/curriculum.types';
 
 export interface CurriculumState {
   curriculums: ICurriculum[];
