+++ client/src/app/state/grades/grades.reducer.ts
@@ -1,6 +1,6 @@
 import { createReducer, on } from '@ngrx/store';
 import * as GradeActions from './grades.actions';
-import { IGrades } from 'app/modules/admin/grades/grades.types';
+import { IGrades } from 'app/modules/superadmin/grades/grades.types';
 
 export interface GradeState {
   gradesByCurriculum: { [curriculumId: string]: IGrades[] };
