+++ client/src/app/state/subjects/subjects.actions.ts
@@ -1,5 +1,5 @@
 import { createAction, props } from '@ngrx/store';
-import { ISubjects } from 'app/modules/admin/subjects/subject.types';
+import { ISubjects } from 'app/modules/superadmin/subjects/subject.types';
 
 export const loadSubjects = createAction('[Subject] Load Subjects', props<{gradeId: string}>());
 export const loadSubjectsSuccess = createAction('[Subject] Load Subjects Success', props<{ gradeId: string, subjects: ISubjects[] }>());
